refactor : 기존 디버그용 클래스들 패키지 이동
